+++ src/main/java/com/example/whales/api/dto/response/LoginResponse.java
@@ -1,0 +1,15 @@
+package com.example.whales.api.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class LoginResponse {
+    private String accesstoken;
+    private String refreshtoken;
+}
