--- src/main/java/com/example/whales/dto/request/SignupRequestDto.java
@@ -1,17 +1,0 @@
-package com.example.whales.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class SignupRequestDto {
-    private String username;
-    private String email;
-    private String password;
-    private String provider;
-}
